requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.23.5
tensorflow-cpu==2.10.0
Pillow==10.2.0
gunicorn==20.1.0
